crud işlemleri devam,uow

+++ core/Coban.Application/Abstractions/Repositories/Base/Read/IAsyncReadRepository.cs
@@ -5,7 +5,7 @@ namespace Coban.Application.Abstractions.Repositories.Base.Read;
 
 public interface IAsyncReadRepository<T> where T : BaseEntity
 {
-    Task<T?> GetByIdAsync(int id, bool tracking = true, CancellationToken cancellationToken = default);
+    Task<T?> GetByIdAsync(long id, bool tracking = true, CancellationToken cancellationToken = default);
     IQueryable<T> GetWhere(Expression<Func<T, bool>> predicate, bool tracking = true);
     IQueryable<T> GetAll(bool tracking = true);
 }
